+++ src/main.js
@@ -1,3 +1,3 @@
 require('./form');
 require('./game');
-require('./reviews');
+require('./reviews-cont/reviews');
